#31-S3와 DB Access 관련 로직을 분리
